backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.1
langchain-google-genai
langchain-text-splitters
chromadb
python-dotenv
tenacity
google-generativeai
requests
gTTS
pypdf
unstructured
